FIX: correction of a lot of errors detected by pycharm
