+++ cmp/stages/connectome/connectome.py
@@ -27,10 +27,10 @@ from nipype.interfaces.base import BaseInterface, BaseInterfaceInputSpec, \
 from nipype.utils.filemanip import split_filename
 
 # Own imports
-from cmtklib.connectome import mrtrixcmat, cmat, prob_cmat, probtrackx_cmat
 from nipype.interfaces.mrtrix3.connectivity import BuildConnectome
 from cmtklib.interfaces.mrtrix3 import FilterTractogram
 from cmp.stages.common import Stage
+from cmtklib.connectome import CMTK_cmat
 
 
 class ConnectomeConfig(HasTraits):
@@ -43,136 +43,6 @@ class ConnectomeConfig(HasTraits):
     log_visualization = Bool(True)
     circular_layout = Bool(False)
     subject = Str
-
-
-class CMTK_cmatInputSpec(BaseInterfaceInputSpec):
-    track_file = InputMultiPath(File(exists=True), desc='Tractography result', mandatory=True)
-    roi_volumes = InputMultiPath(File(exists=True), desc='ROI volumes registered to diffusion space')
-    parcellation_scheme = traits.Enum('Lausanne2008', ['Lausanne2008', 'Lausanne2018', 'NativeFreesurfer', 'Custom'],
-                                      usedefault=True)
-    roi_graphMLs = InputMultiPath(File(exists=True), desc='GraphML description of ROI volumes (Lausanne2018)')
-    atlas_info = Dict(mandatory=False, desc="custom atlas information")
-    compute_curvature = traits.Bool(True, desc='Compute curvature', usedefault=True)
-    additional_maps = traits.List(File, desc='Additional calculated maps (ADC, gFA, ...)')
-    output_types = traits.List(Str, desc='Output types of the connectivity matrices')
-    probtrackx = traits.Bool(False)
-    voxel_connectivity = InputMultiPath(File(exists=True),
-                                        desc="ProbtrackX connectivity matrices (# seed voxels x # target ROIs)")
-
-
-class CMTK_cmatOutputSpec(TraitedSpec):
-    endpoints_file = File()
-    endpoints_mm_file = File()
-    final_fiberslength_files = OutputMultiPath(File())
-    filtered_fiberslabel_files = OutputMultiPath(File())
-    final_fiberlabels_files = OutputMultiPath(File())
-    streamline_final_file = File()
-    connectivity_matrices = OutputMultiPath(File())
-
-
-class CMTK_cmat(BaseInterface):
-    input_spec = CMTK_cmatInputSpec
-    output_spec = CMTK_cmatOutputSpec
-    
-    def _run_interface(self, runtime):
-        if isdefined(self.inputs.additional_maps):
-            additional_maps = dict(
-                (split_filename(add_map)[1], add_map) for add_map in self.inputs.additional_maps if add_map != '')
-        else:
-            additional_maps = {}
-        
-        if self.inputs.probtrackx:
-            probtrackx_cmat(voxel_connectivity_files=self.inputs.track_file, roi_volumes=self.inputs.roi_volumes,
-                            parcellation_scheme=self.inputs.parcellation_scheme, atlas_info=self.inputs.atlas_info,
-                            output_types=self.inputs.output_types)
-        elif len(self.inputs.track_file) > 1:
-            prob_cmat(intrk=self.inputs.track_file, roi_volumes=self.inputs.roi_volumes,
-                      parcellation_scheme=self.inputs.parcellation_scheme, atlas_info=self.inputs.atlas_info,
-                      output_types=self.inputs.output_types)
-        else:
-            cmat(intrk=self.inputs.track_file[0], roi_volumes=self.inputs.roi_volumes,
-                 roi_graphmls=self.inputs.roi_graphMLs,
-                 parcellation_scheme=self.inputs.parcellation_scheme, atlas_info=self.inputs.atlas_info,
-                 compute_curvature=self.inputs.compute_curvature,
-                 additional_maps=additional_maps, output_types=self.inputs.output_types)
-            
-            if 'cff' in self.inputs.output_types:
-                cvt = cmtk.CFFConverter()
-                cvt.inputs.title = 'Connectome mapper'
-                cvt.inputs.nifti_volumes = self.inputs.roi_volumes
-                cvt.inputs.tract_files = ['streamline_final.trk']
-                cvt.inputs.gpickled_networks = glob.glob(os.path.abspath("connectome_*.gpickle"))
-                cvt.run()
-        
-        return runtime
-    
-    def _list_outputs(self):
-        outputs = self._outputs().get()
-        outputs['endpoints_file'] = os.path.abspath('endpoints.npy')
-        outputs['endpoints_mm_file'] = os.path.abspath('endpointsmm.npy')
-        outputs['final_fiberslength_files'] = glob.glob(os.path.abspath('final_fiberslength*'))
-        outputs['filtered_fiberslabel_files'] = glob.glob(os.path.abspath('filtered_fiberslabel*'))
-        outputs['final_fiberlabels_files'] = glob.glob(os.path.abspath('final_fiberlabels*'))
-        outputs['streamline_final_file'] = os.path.abspath('streamline_final.trk')
-        outputs['connectivity_matrices'] = glob.glob(os.path.abspath('connectome*'))
-        
-        return outputs
-
-
-class CMTK_mrtrixcmatInputSpec(BaseInterfaceInputSpec):
-    track_file = InputMultiPath(File(exists=True), desc='Tractography result', mandatory=True)
-    fod_file = InputMultiPath(File(exists=True),
-                              desc='Input image containing the spherical harmonics of the fibre orientation distributions',
-                              mandatory=True)
-    roi_volumes = InputMultiPath(File(exists=True), desc='ROI volumes ')
-    parcellation_scheme = traits.Enum('Lausanne2008', ['Lausanne2008', 'NativeFreesurfer', 'Custom'], usedefault=True)
-    atlas_info = Dict(mandatory=False, desc="custom atlas information")
-    compute_curvature = traits.Bool(True, desc='Compute curvature', usedefault=True)
-    additional_maps = traits.List(File, desc='Additional calculated maps (ADC, gFA, ...)')
-    output_types = traits.List(Str, desc='Output types of the connectivity matrices')
-    # probtrackx = traits.Bool(False)
-    # voxel_connectivity = InputMultiPath(File(exists=True),desc = "ProbtrackX connectivity matrices (# seed voxels x # target ROIs)")
-
-
-class CMTK_mrtrixcmatOutputSpec(TraitedSpec):
-    # endpoints_file = File()
-    # endpoints_mm_file = File()
-    # final_fiberslength_files = OutputMultiPath(File())
-    # filtered_fiberslabel_files = OutputMultiPath(File())
-    # final_fiberlabels_files = OutputMultiPath(File())
-    streamline_final_file = File()
-    connectivity_matrices = OutputMultiPath(File())
-
-
-class CMTK_mrtrixcmat(BaseInterface):
-    input_spec = CMTK_mrtrixcmatInputSpec
-    output_spec = CMTK_mrtrixcmatOutputSpec
-    
-    def _run_interface(self, runtime):
-        if isdefined(self.inputs.additional_maps):
-            additional_maps = dict(
-                (split_filename(add_map)[1], add_map) for add_map in self.inputs.additional_maps if add_map != '')
-        else:
-            additional_maps = {}
-        
-        mrtrixcmat(intck=self.inputs.track_file[0], fod_file=self.inputs.fod_file, roi_volumes=self.inputs.roi_volumes,
-                   parcellation_scheme=self.inputs.parcellation_scheme, atlas_info=self.inputs.atlas_info,
-                   compute_curvature=self.inputs.compute_curvature, additional_maps=additional_maps,
-                   output_types=self.inputs.output_types)
-        
-        return runtime
-    
-    def _list_outputs(self):
-        outputs = self._outputs().get()
-        # outputs['endpoints_file'] = os.path.abspath('endpoints.npy')
-        # outputs['endpoints_mm_file'] = os.path.abspath('endpointsmm.npy')
-        # outputs['final_fiberslength_files'] = glob.glob(os.path.abspath('final_fiberslength*'))
-        # outputs['filtered_fiberslabel_files'] = glob.glob(os.path.abspath('filtered_fiberslabel*'))
-        # outputs['final_fiberlabels_files'] = glob.glob(os.path.abspath('final_fiberlabels*'))
-        outputs['streamline_final_file'] = os.path.abspath('streamline_final.tck')
-        outputs['connectivity_matrices'] = glob.glob(os.path.abspath('connectome*'))
-        
-        return outputs
 
 
 class ConnectomeStage(Stage):
@@ -296,3 +166,59 @@ class ConnectomeStage(Stage):
     
     def has_run(self):
         return os.path.exists(os.path.join(self.stage_dir, "compute_matrice", "result_compute_matrice.pklz"))
+
+# # OLD
+# class CMTK_mrtrixcmatInputSpec(BaseInterfaceInputSpec):
+#     track_file = InputMultiPath(File(exists=True), desc='Tractography result', mandatory=True)
+#     fod_file = InputMultiPath(File(exists=True),
+#                               desc='Input image containing the spherical harmonics of the fibre orientation distributions',
+#                               mandatory=True)
+#     roi_volumes = InputMultiPath(File(exists=True), desc='ROI volumes ')
+#     parcellation_scheme = traits.Enum('Lausanne2008', ['Lausanne2008', 'NativeFreesurfer', 'Custom'], usedefault=True)
+#     atlas_info = Dict(mandatory=False, desc="custom atlas information")
+#     compute_curvature = traits.Bool(True, desc='Compute curvature', usedefault=True)
+#     additional_maps = traits.List(File, desc='Additional calculated maps (ADC, gFA, ...)')
+#     output_types = traits.List(Str, desc='Output types of the connectivity matrices')
+#     # probtrackx = traits.Bool(False)
+#     # voxel_connectivity = InputMultiPath(File(exists=True),desc = "ProbtrackX connectivity matrices (# seed voxels x # target ROIs)")
+
+
+# class CMTK_mrtrixcmatOutputSpec(TraitedSpec):
+#     # endpoints_file = File()
+#     # endpoints_mm_file = File()
+#     # final_fiberslength_files = OutputMultiPath(File())
+#     # filtered_fiberslabel_files = OutputMultiPath(File())
+#     # final_fiberlabels_files = OutputMultiPath(File())
+#     streamline_final_file = File()
+#     connectivity_matrices = OutputMultiPath(File())
+
+
+# class CMTK_mrtrixcmat(BaseInterface):
+#     input_spec = CMTK_mrtrixcmatInputSpec
+#     output_spec = CMTK_mrtrixcmatOutputSpec
+    
+#     def _run_interface(self, runtime):
+#         if isdefined(self.inputs.additional_maps):
+#             additional_maps = dict(
+#                 (split_filename(add_map)[1], add_map) for add_map in self.inputs.additional_maps if add_map != '')
+#         else:
+#             additional_maps = {}
+        
+#         mrtrixcmat(intck=self.inputs.track_file[0], fod_file=self.inputs.fod_file, roi_volumes=self.inputs.roi_volumes,
+#                    parcellation_scheme=self.inputs.parcellation_scheme, atlas_info=self.inputs.atlas_info,
+#                    compute_curvature=self.inputs.compute_curvature, additional_maps=additional_maps,
+#                    output_types=self.inputs.output_types)
+        
+#         return runtime
+    
+#     def _list_outputs(self):
+#         outputs = self._outputs().get()
+#         # outputs['endpoints_file'] = os.path.abspath('endpoints.npy')
+#         # outputs['endpoints_mm_file'] = os.path.abspath('endpointsmm.npy')
+#         # outputs['final_fiberslength_files'] = glob.glob(os.path.abspath('final_fiberslength*'))
+#         # outputs['filtered_fiberslabel_files'] = glob.glob(os.path.abspath('filtered_fiberslabel*'))
+#         # outputs['final_fiberlabels_files'] = glob.glob(os.path.abspath('final_fiberlabels*'))
+#         outputs['streamline_final_file'] = os.path.abspath('streamline_final.tck')
+#         outputs['connectivity_matrices'] = glob.glob(os.path.abspath('connectome*'))
+        
+#         return outputs
